pyproject.toml
[tool.poetry]
name = "pjecz-citas-v2-admin"
version = "2.1.0"
description = "Citas V2 admin es un sistema web hecho con Flask."
authors = ["Guillermo Valdes <guillermo@movimientolibre.com>"]
license = "AGPL3"
readme = "README.md"
packages = [{include = "citas_admin"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
cryptography = "^42.0.5"
email-validator = "^2.1.1"
flask = "^3.0.2"
flask-login = "^0.6.3"
flask-moment = "^1.0.5"
flask-sqlalchemy = "^3.1.1"
flask-socketio = "^5.3.6"
flask-wtf = "^1.2.1"
google-auth = "^2.29.0"
google-cloud-secret-manager = "^2.19.0"
google-cloud-storage = "^2.16.0"
google-cloud = "^0.34.0"
google-generativeai = "^0.4.1"
gunicorn = "^21.2.0"
hashids = "^1.3.1"
jinja2 = "^3.1.3"
openpyxl = "^3.1.2"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
rq = "^1.16.1"
sqlalchemy = "^2.0.29"
sqlalchemy-utils = "^0.41.2"
unidecode = "^1.3.8"
uvicorn = "^0.29.0"
werkzeug = "^3.0.1"
wtforms = "^3.1.2"
quill-delta = {extras = ["html"], version = "^1.0.3"}
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.1.1"
pylint = "^3.1.0"
pylint-sqlalchemy = "^0.3.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128

[tool.isort]
line_length = 128
profile = "black"
